Add SigLIP weights
